Git Status language change
I updated Git a while ago and noticed the message from Git changed and it perpetually showed my branches as dirty.

The language was changed in Git in this commit:
https://github.com/git/git/commit/2a0e6cdedab306eccbd297c051035c13d0266343

This fixes the bash prompt to look for the new language. Let me know if you want it to keep looking for the old language too.